--- a/compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
+++ b/compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -33,7 +33,11 @@ static llvm::cl::opt<bool> clCheckIRBeforeLLVMConversion(
     "iree-codegen-check-ir-before-llvm-conversion",
     llvm::cl::desc("Runs the pass to check the IR generated from LLVMCPU "
                    "before conversion to LLVM IR"),
-    llvm::cl::init(true));
+    llvm::cl::init(false));
+
+static llvm::cl::opt<bool> clEnableHoistPadding(
+    "iree-llvmcpu-enable-hoist-padding",
+    llvm::cl::desc("Flag to enable hoist padding"), llvm::cl::init(false));
 
 //===---------------------------------------------------------------------===//
 // Default allocation functions for CPU backend
@@ -245,10 +249,12 @@ void addDoubleTilingPadExpertPassPipeline(OpPassManager &passManager) {
     passManager.addNestedPass<func::FuncOp>(createCSEPass());
   }
 
-  auto pad = [&](std::string anchorOpName, ArrayRef<int64_t> paddingDims) {
+  auto pad = [&](std::string anchorOpName, ArrayRef<int64_t> paddingDims,
+                 ArrayRef<int64_t> packPaddings = {}) {
     LinalgFusePassOptions options;
     options.pad = true;
     options.anchorOpName = anchorOpName;
+    options.packPaddings.assign(packPaddings.begin(), packPaddings.end());
     options.paddingDimensions.assign(paddingDims.begin(), paddingDims.end());
     passManager.addNestedPass<func::FuncOp>(createLinalgFusePass(options));
   };
@@ -270,20 +276,17 @@ void addDoubleTilingPadExpertPassPipeline(OpPassManager &passManager) {
   }
 
   paddingDims = {2};
-  pad("linalg.matmul", paddingDims);
+  pad("linalg.matmul", paddingDims, {1, 1, 0});
 
-  // TODO(hanchung): Evaluate if we want to enable hoist paddings.
-  //{
-  // LinalgFusePassOptions options;
-  // options.pad = true;
-  // options.anchorOpName = "linalg.matmul";
-  // options.hoistPaddings = SmallVector<int64_t>{2, 3, 0};
-  // options.transposePaddings = SmallVector<std::string>{"0:1", "0:1",
-  // "0:1"};
-  // passManager.addNestedPass<func::FuncOp>(createLinalgFusePass(options));
-  // passManager.addNestedPass<func::FuncOp>(createCanonicalizerPass());
-  // passManager.addNestedPass<func::FuncOp>(createCSEPass());
-  //}
+  if (clEnableHoistPadding) {
+    LinalgFusePassOptions options;
+    options.pad = true;
+    options.anchorOpName = "linalg.matmul";
+    options.hoistPaddings = SmallVector<int64_t>{2, 3, 0};
+    passManager.addNestedPass<func::FuncOp>(createLinalgFusePass(options));
+    passManager.addNestedPass<func::FuncOp>(createCanonicalizerPass());
+    passManager.addNestedPass<func::FuncOp>(createCSEPass());
+  }
 
   // Fold dim(pad) away before vectorization.
   passManager.addPass(memref::createResolveShapedTypeResultDimsPass());
@@ -333,9 +336,6 @@ void addDoubleTilingExpertPassPipeline(OpPassManager &passManager,
 
   // Add the sandbox single tiling expert to tile and vectorize.
   {
-    // The options are derived from sandbox codegen driver. hoistPadding options
-    // does not work in IREE cases. It's fine to not have it, since it's already
-    // generating the IR as same as sandbox.
     LinalgSingleTilingExpertPassOptions options;
     options.vectorize = true;
     options.tilingLevel =
@@ -465,7 +465,9 @@ static void addLowerToLLVMPasses(OpPassManager &passManager) {
   passManager.addNestedPass<func::FuncOp>(createPolynomialApproximationPass());
 
   // Checking stack allocation before converting to CF dialect is easier.
-  if (clCheckIRBeforeLLVMConversion) {
+  // Do not check allocation if hoist-padding is enabled. It intends to allocate
+  // big stack buffers for better accessing.
+  if (clCheckIRBeforeLLVMConversion && !clEnableHoistPadding) {
     passManager.addPass(createLLVMCPUCheckIRBeforeLLVMConversionPass());
   }