--- a/compiler/src/iree/compiler/Codegen/Sandbox/LinalgTensorCodegenDriver.cpp
+++ b/compiler/src/iree/compiler/Codegen/Sandbox/LinalgTensorCodegenDriver.cpp
@@ -297,8 +297,12 @@ void LinalgFusePass::runOnOperation() {
       .vectorizeIf(vectorize, "", nullptr, vectorizePadding);
 
   // Created a nested OpPassManager and run.
+  // XXX(hanchung): Do not run LinalgStrategyEnablePass when padding is enabled.
+  // This aligns the Sandbox behavior. Otherwise, CSE and canonicalizer will
+  // mess up fold/nofold pad ops.
+  // Note that Sandbox uses dialect approach, so there are slight difference.
   OpPassManager dynamicPM(func::FuncOp::getOperationName());
-  strategy.configurePassPipeline(dynamicPM, funcOp.getContext());
+  strategy.configurePassPipeline(dynamicPM, funcOp.getContext(), !pad);
 
   if (failed(runPipeline(dynamicPM, funcOp))) {
     return signalPassFailure();