This is only available through flags because we may not want huge stack
allocation. It needs more exploration about if it is a good approach.
Adding an option to IREE to enable more performance work.