@hanhanW 2022-04-28
To repro:

Use https://gist.githubusercontent.com/hanhanW/56fd8e37e2e0b1dfbe88855d967a8c61/raw

Run:

- `iree-translate -iree-mlir-to-vm-bytecode-module -iree-hal-target-backends=dylib-llvm-aot --iree-llvm-link-embedded=true --iree-llvm-target-triple=x86_64-pc-linux-gnu --iree-llvm-target-cpu-features=host ~/matmul_bias_add.mlir -o /tmp/b.vmfb`
- `iree-benchmark-module --module_file=/tmp/b.vmfb --driver=dylib-sync --entry_function=matmul_bias_add --function_input=64x1536xf32 --function_input=1536x64xf32 --function_input=64xf32`

IREE IR before bufferization: https://gist.githubusercontent.com/hanhanW/b623547c0572619cfa5730a2fe6b2633/raw
Sandbox IR before bufferization: https://gist.githubusercontent.com/hanhanW/b2a74b07fb88c13c48c81f3a14db550e/raw