@hanhanW 2022-04-28
One of issues is that two `memref<2x96x16x32xf32>` are created and unnecessary copies are created. The I'll file a bug.