@hanhanW 2022-05-13
The perf in sandbox and IREE are aligned. I benchmarked three shapes (from MiniLM) Bert models (on A100).

Single-threaded:

| Shape (NxMxK)  | Before  | After   |
| ---------------- | ------- | ------- |
| 128x384x384        | 0.301 ms  | 0.384 ms  |
| 128x384x1536  | 1.43 ms  | 1.83 ms  |
| 128x1536x384      | 1.93 ms  | 1.45 ms |

Multi-threaded:

| Shape (NxMxK)  | Before  | After   |
| ---------------- | ------- | ------- |
| 128x384x384        | 0.178 ms  | 0.163 ms  |
| 128x384x1536  | 0.343 ms  | 0.473 ms  |
| 128x1536x384      | 0.396 ms  | 0.501 ms |

Looks like it only benefits the `128x1536x384` case on single-threaded. We may want to tune it a bit more for multi-threaded though.